Move model files and utils
