+++ tabarena/tabarena/models/limix/_vendor/__init__.py
@@ -3,7 +3,7 @@
 Upstream: https://github.com/limix-ldm-ai/LimiX (Apache-2.0).
 The LimiX repo is not packaged for pip, so the inference-path sources are
 vendored here. Imports were rewritten to live under
-``tabarena.benchmark.models.ag.limix._vendor``.
+``tabarena.models.limix._vendor``.
 """
 
 from __future__ import annotations
