+++ tabarena/tabarena/models/limix/info.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from tabarena.benchmark.models.ag.limix.limix_model import LimiXModel
+from tabarena.models.limix.model import LimiXModel
 from tabarena.models._model_info import ModelInfo
 from tabarena.models.limix.hpo import gen_limix
 from tabarena.nips2025_utils.artifacts.method_metadata import MethodMetadata
